Split generate_image.py; series overline; standings clinch boxes
- Refactor 2656-line generate_image.py into image_assets.py,
  image_utils.py, image_standings.py, image_box.py; generate_image.py
  becomes a thin orchestration layer with re-exports for compat
- Series score overline: regular season shows on final game of set
  (game N of N); postseason shows only on clinch/sweep
- Standings sidebar: draw box around logo when division clinched (1px)
  or home-field advantage clinched (2px)
- Fetch and store clinchIndicator from MLB standings API
- Update demo to show no-hitter, postponed, and sub_event boxes;
  simplify to single panel
